move consts to common folder
