+++ components/youtube_data_api/actions/update-playlist/common.mjs
@@ -1,4 +1,4 @@
-import consts from "../../consts.mjs";
+import consts from "../../common/consts.mjs";
 
 export default {
   methods: {
